Add random colored borders to browsers and targets that are sufficiently bright, and add function to pass them to GUI

+++ modules/skybrowser/include/screenspaceskytarget.h
@@ -59,7 +59,7 @@ namespace openspace {
         GLuint _vertexBuffer = 0;
         float _fieldOfView = 100.f;
         ScreenSpaceSkyBrowser* _skyBrowser;
-        glm::vec3 _borderColor;
+        glm::ivec3 _borderColor;
     };
 }
